Version5
Successfully incorporated sending of OTP to Email and mobile number in the previous version (i.e version 4)
The API attached for sending SMS is 2factor which is a third-party SMS vendor.
Shortcomings: The Loading time after submission of mobile number and Email Id is way too long. We will try to fix it in the next version probably.